feat: sub-dir option is now possible
note: had a stupid error that took some debugging time to solve
in storage.go:uniqueDstPath function.

It was literally joining the file name onto itself, because I wrote a stupid code.
looking like: <some_path>/pdf/abcd.pdf/abcd_1.pdf

That was not my only stupidity. I wrote the break in the for loop only and if only
		 the error value is "file does not exist".

I did not check if there could be any other errors. On a file system. lol